errorPage + validation

+++ src/main/java/com/aimprosoft/sandbox/controller/action/post/department/DeleteDepartment.java
@@ -3,6 +3,7 @@ package com.aimprosoft.sandbox.controller.action.post.department;
 import com.aimprosoft.sandbox.controller.action.Action;
 import com.aimprosoft.sandbox.dao.impl.DepartmentDAO;
 import com.aimprosoft.sandbox.dao.impl.EmployeeDAO;
+import com.aimprosoft.sandbox.util.validator.Validator;
 import org.apache.log4j.Logger;
 
 import javax.servlet.http.HttpServletRequest;
@@ -17,9 +18,14 @@ public class DeleteDepartment implements Action {
 
     @Override
     public void execute(HttpServletRequest request, HttpServletResponse response, EmployeeDAO employeeDAO, DepartmentDAO departmentDAO) throws IOException {
-        Long departmentId = Long.parseLong(request.getParameter("id"));
-        departmentDAO.deleteDepartmentById(departmentId);
-        LOG.info("Department " + request.getParameter("name") + " was removed!");
-        response.sendRedirect("/");
+        String id = request.getParameter("id");
+        if (Validator.validateId(id)) {
+            Long departmentId = Long.parseLong(id);
+            departmentDAO.deleteDepartmentById(departmentId);
+            LOG.info("Department " + request.getParameter("name") + " was removed!");
+            response.sendRedirect("/");
+        } else {
+            response.sendRedirect("?action-get=error");
+        }
     }
 }
